+++ R/colplot.R
@@ -41,7 +41,7 @@
 #' @export
 #' @examples {
 #' a <- ef_cetesb("CO", c("PC_G", "PC_FE", "PC_FG"), agemax = 20)
-#' colplot(df = a, ylab = "CO [g/km]", theme = "dark")
+#' colplot(df = a, ylab = "CO [g/km]", theme = "dark", pch = NULL, type = "l")
 #' colplot(df = a, cols = "PC_FG", main = "EF", ylab = "CO [g/km]")
 #' colplot(df = a, ylab = "CO [g/km]", theme = "clean")
 #' }